+++ src/main/java/com/grinner/game/jnbjt/pojo/vo/AssetVO.java
@@ -2,6 +2,7 @@ package com.grinner.game.jnbjt.pojo.vo;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -17,7 +18,7 @@ public class AssetVO {
 
     private Integer id;
     //每份资产对应的“人时”
-    private Integer value;
+    private Double value;
 
     protected String name;
 }
